Reorganizing code to separate DocumentTracker from Package code.
