+++ VSDocumentReopen/VSDocumentReopen/ReopenPackage.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.InteropServices;
@@ -29,8 +28,6 @@ namespace VSDocumentReopen
 		public const string PackageGuidString = "b30147a1-6fbc-4b94-bf01-123d837c4fe2";
 
 		private readonly DTE2 _dte;
-		private readonly SolutionEvents _solutionEvents;
-		private readonly DocumentEvents _documentEvents;
 
 		/// <summary>
 		/// Initializes a new instance of the <see cref="Reopen"/> class.
@@ -38,9 +35,6 @@ namespace VSDocumentReopen
 		public ReopenPackage()
 		{
 			_dte = GetGlobalService(typeof(DTE)) as DTE2 ?? throw new NullReferenceException($"Unable to get service {nameof(DTE2)}");
-
-			_solutionEvents = _dte.Events.SolutionEvents;
-			_documentEvents = _dte.Events.DocumentEvents;
 		}
 
 		protected override async Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
@@ -56,25 +50,7 @@ namespace VSDocumentReopen
 
 			EnforceKeyBinding();
 
-			_solutionEvents.Opened += () =>
-			{
-				DocumentTracker.Instance.Clear();
-				_documentEvents.DocumentClosing += DocumentEventsOnDocumentClosing;
-
-				Debug.WriteLine("Solution opened");
-			};
-			_solutionEvents.AfterClosing += () =>
-			{
-				_documentEvents.DocumentClosing -= DocumentEventsOnDocumentClosing;
-				DocumentTracker.Instance.Clear();
-
-				Debug.WriteLine("Solution closed");
-			};
-		}
-
-		private void DocumentEventsOnDocumentClosing(Document document)
-		{
-			DocumentTracker.Instance.AddClosed(document);
+			new DocumentTracker(_dte);
 		}
 
 		/// <summary>
